Use strdisplaywidth where available
In vim 7.3, the strdisplaywidth function was added to give an accurate
count of the number of screen columns that will be taken up by a given
string.  Tabular was written before vim 7.3, and was implementing this
functionality itself (at least for tabs; it wasn't trying to handle
doublewide characters).

This changeset allows us to take advantage of strdisplaywidth where it
exists.