Don't pad after 1st field if it's always empty
Further special case the 1st field of a line so that if it is blank for
all lines, the padding after it won't be added.  This prevents another
case of padding creep, where each run through would add (the default)
space after the empty field, and the next run through would preserve
that space as leading indent, and add a new one after the first field
(which would be blank again, because leading indent had been stripped).